+++ RELEASES.md
@@ -142,3 +142,12 @@ Based on:
 - [Go v0.4.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.4.0 - go-client-sdk
 - [PyPI v1.2.0] https://pypi.org/project/brease-sdk/1.2.0 - python-client-sdk
 - [NPM v0.2.0] https://www.npmjs.com/package/brease-sdk/v/0.2.0 - typescript-client-sdk
+
+## 2023-05-16 00:41:34
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.32.0 (2.28.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.5.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.5.0 - go-client-sdk
+- [NPM v0.2.0] https://www.npmjs.com/package/brease-sdk/v/0.2.0 - typescript-client-sdk